+++ packages/atomic/src/components/commerce/index.ts
@@ -8,6 +8,7 @@ export {AtomicCommerceNoProducts} from './atomic-commerce-no-products/atomic-com
 export {AtomicCommercePager} from './atomic-commerce-pager/atomic-commerce-pager.js';
 export {AtomicCommerceProductList} from './atomic-commerce-product-list/atomic-commerce-product-list.js';
 export {AtomicCommerceProductsPerPage} from './atomic-commerce-products-per-page/atomic-commerce-products-per-page.js';
+export {AtomicCommerceQueryError} from './atomic-commerce-query-error/atomic-commerce-query-error.js';
 export {AtomicCommerceQuerySummary} from './atomic-commerce-query-summary/atomic-commerce-query-summary.js';
 export {AtomicCommerceRecommendationInterface} from './atomic-commerce-recommendation-interface/atomic-commerce-recommendation-interface.js';
 export {AtomicCommerceRecommendationList} from './atomic-commerce-recommendation-list/atomic-commerce-recommendation-list.js';
@@ -20,6 +21,6 @@ export {AtomicCommerceSearchBoxRecentQueries} from './atomic-commerce-search-box
 export {AtomicCommerceSortDropdown} from './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown.js';
 export {AtomicCommerceText} from './atomic-commerce-text/atomic-commerce-text.js';
 export {AtomicProduct} from './atomic-product/atomic-product.js';
-export {AtomicProductExcerpt} from './atomic-product-excerpt/atomic-product-excerpt.js';
+export {AtomicProductChildren} from './atomic-product-children/atomic-product-children.js';
 export {AtomicProductTemplate} from './atomic-product-template/atomic-product-template.js';
 export {AtomicProductText} from './atomic-product-text/atomic-product-text.js';